music pro/requirements.txt
opencv-python
mediapipe
pyautogui
google-generativeai